renamed offset variables in IndexReader and removed unnecessary cast to long in IndexEntry crc getter

+++ src/main/java/tech/sourced/siva/IndexEntry.java
@@ -72,7 +72,7 @@ public class IndexEntry extends Header {
      * @return CRC32.
      */
     public final long getCrc32() {
-        return (long) this.crc32;
+        return this.crc32;
     }
 
 }
